requirements.txt
Flask
numpy
pandas
scikit-learn
bcrypt
python-dotenv